Fixing test_db_task.py third attempt

+++ tests/test_db_task.py
@@ -44,7 +44,11 @@ class Evaluate(unittest.TestCase):
     def test_delete(self):
         data = {"_id": 2, "First name": "Andres"}
         mongo = MongoDbConnector("localhost,27017,db_task,test", ["2", data])
-        actual = mongo.delete()
+        document = mongo.select_by_id()
+        if document is not None:
+            actual = mongo.delete()
+        else:
+            actual = "Document doesn't exist"
         expected = {"response": "Document successfully deleted"}
         self.assertEqual(expected, actual)
